make more things bytes

+++ src/offset/group8/MemoryEfficientPoint.java
@@ -3,8 +3,8 @@ package offset.group8;
 public class MemoryEfficientPoint {
     public byte x;
     public byte y;
-    public int value;
-    public int owner;
+    public byte value;
+    public byte owner;
     public boolean change;
 
     public MemoryEfficientPoint() { x = 0; y = 0; value =1; owner = -1;}
@@ -12,8 +12,8 @@ public class MemoryEfficientPoint {
     public MemoryEfficientPoint(int xx, int yy, int va, int ow) {
         x = (byte) xx;
         y = (byte) yy;
-        this.value = va;
-        this.owner = ow;
+        this.value = (byte) va;
+        this.owner = (byte) ow;
         this.change = false;
     }
 }
